Controller: fix JSONObject, time_stamp
1. We use append in JSONObject, but it should be put.
2. time_stamp of event should be minute, so divide it by 60.

+++ src/main/java/Controller/Problem.java
@@ -7,9 +7,9 @@ public class Problem {
     private final String id;
     private int timeLimit;
     private int memoryLimit;
-    private long testDataTimeStamp;
+    private int testDataTimeStamp;
 
-    public Problem(String id, int timeLimit, int memoryLimit, long testDataTimeStamp) {
+    public Problem(String id, int timeLimit, int memoryLimit, int testDataTimeStamp) {
         this.id = id;
         this.timeLimit = timeLimit;
         this.memoryLimit = memoryLimit;
@@ -24,7 +24,7 @@ public class Problem {
         return this.memoryLimit;
     }
 
-    public long getTestDataTimeStamp() {
+    public int getTestDataTimeStamp() {
         return this.testDataTimeStamp;
     }
 }
